+++ src/main/java/com/chunjae/privatepro05/biz/UserServiceImpl.java
@@ -59,7 +59,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public Kuser getUserByPk(Integer id) {
+    public Kuser getUserByPk(Long id) {
         return userMapper.getUserByPk(id);
     }
